+++ Assets/Anims.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d66cc75cc40144fd3be9e9c9af6a5e7a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
