+++ Assets/Anims/Mobile - Flappy Bird - Version 12 Sprites_71 1.controller.meta
