+++ Assets/Scripts/BirdMovement.cs
@@ -1,0 +1,31 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class BirdMovement : MonoBehaviour
+{
+    public GameManager gameManager;
+    public float velocity = 1;
+    private Rigidbody2D rb;
+    // Start is called before the first frame update
+    void Start()
+    {
+        rb = GetComponent<Rigidbody2D>();
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        if (Input.GetMouseButtonDown(0))
+        {
+            rb.velocity = Vector2.up*velocity;
+        }
+    }
+    private void OnCollisionEnter2D(Collision2D collision)
+    {
+        gameManager.GameOver();
+    }
+    void OnTriggerEnter2D(Collider2D col){
+        gameManager.addScore();
+    }
+}
