+++ Assets/Scripts/BirdMovement.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2a1e489e1cf657144b62c95e487246a5
+guid: 31647008df4b34050b95a405c99f1b32
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
