+++ Assets/Scripts/PipeMove.cs
@@ -1,0 +1,14 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PipeMove : MonoBehaviour
+{
+     public float speed = 1;
+    // Update is called once per frame
+    void Update()
+    {
+        transform.position += Vector3.left*speed*Time.deltaTime;
+    }
+
+}
