+++ Assets/Sprites/Mobile - Flappy Bird - Version 12 Sprites.png.meta
@@ -1730,9 +1730,9 @@ TextureImporter:
       Mobile - Flappy Bird - Version 12 Sprites_54: -1460865571
       Mobile - Flappy Bird - Version 12 Sprites_66: -578743416
       Mobile - Flappy Bird - Version 12 Sprites_61: -2095331823
-      Mobile - Flappy Bird - Version 12 Sprites_1: -1618468559
       Mobile - Flappy Bird - Version 12 Sprites_19: 1108703421
-      Mobile - Flappy Bird - Version 12 Sprites_32: -527416111
+      Mobile - Flappy Bird - Version 12 Sprites_1: -1618468559
+      Mobile - Flappy Bird - Version 12 Sprites_44: -973708966
       Mobile - Flappy Bird - Version 12 Sprites_60: 2121295935
       Mobile - Flappy Bird - Version 12 Sprites_52: 19636039
       Mobile - Flappy Bird - Version 12 Sprites_10: 1907504204
@@ -1740,7 +1740,7 @@ TextureImporter:
       Mobile - Flappy Bird - Version 12 Sprites_26: 932101280
       Mobile - Flappy Bird - Version 12 Sprites_46: -1188653768
       Mobile - Flappy Bird - Version 12 Sprites_41: 1648134618
-      Mobile - Flappy Bird - Version 12 Sprites_44: -973708966
+      Mobile - Flappy Bird - Version 12 Sprites_32: -527416111
       Mobile - Flappy Bird - Version 12 Sprites_34: 1238000314
       Mobile - Flappy Bird - Version 12 Sprites_63: 1507006268
       Mobile - Flappy Bird - Version 12 Sprites_23: 795588114
