--- Assets/move.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 23bb847f3673bc14393b25512f7a7aee
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
